convert Agent to TS
